VLM_damage_recognition/requirements.txt
torch>=2.0.0
transformers>=4.35.0
Pillow>=10.0.0
piexif>=1.1.3
rasterio>=1.3.0
requests>=2.31.0
pydantic>=2.0.0
python-dotenv>=1.0.0
numpy>=1.24.0
scipy>=1.10.0
